feat: add synapse.storage and StorageContext (#153)
Replaces createStorage()

* synapse.storage.upload() and synapse.storage.download() for easy-path
* StorageContext contains a relationship with a specific SP + data set
* synapse.storage.createContext() to manage them directly
* context.download() and context.upload() (etc.) for dealing directly with
  that SP + data set

Closes: https://github.com/FilOzone/synapse-sdk/issues/135
