+++ .gitignore
@@ -4,3 +4,5 @@ package-lock.json
 yarn.lock
 bun.lock
 .zed
+FilOzone-*/
+filecoin-project-*/
