+++ src/storage/index.ts
@@ -1,2 +1,4 @@
 // Export storage components
 export { StorageService } from './service.js'
+export { StorageContext } from './context.js'
+export { StorageManager } from './manager.js'
