+++ utils/example-storage-simple.js
@@ -8,6 +8,9 @@
 /**
  * Simple Storage Example - Minimal upload/download demonstration
  *
+ * This example shows the simplest way to use Synapse SDK's storage API.
+ * The SDK automatically handles provider selection and data set creation.
+ *
  * Usage:
  *   PRIVATE_KEY=0x... WARM_STORAGE_ADDRESS=0x... node example-storage-simple.js
  */
@@ -39,24 +42,26 @@ async function main () {
 
   console.log('Connected to:', RPC_URL)
 
-  // Create storage service
-  const storage = await synapse.createStorage()
-  console.log('Service provider:', storage.serviceProvider)
-  console.log('Data set ID:', storage.dataSetId)
+  // The synapse.storage API auto-manages contexts for you
+  // No need to explicitly create a storage context unless you need specific control
+  console.log('Storage API ready. Will auto-select provider on first upload.')
 
   // Create test data (must be at least 65 bytes for PieceCID calculation)
   const testMessage = 'Hello, Filecoin storage! This message is at least 65 bytes long to meet the minimum requirement for PieceCID calculation.\n'
   const testData = new TextEncoder().encode(testMessage)
   console.log(`\nUploading test data (${testData.length} bytes)...`)
 
-  // Upload
-  const result = await storage.upload(testData)
+  // Upload - the SDK will automatically:
+  // 1. Select a provider
+  // 2. Create or reuse a data set
+  // 3. Upload the data
+  const result = await synapse.storage.upload(testData)
   console.log('Upload complete!')
   console.log('PieceCID:', result.pieceCid)
 
-  // Download
+  // Download - finds any provider with the piece
   console.log('\nDownloading...')
-  const downloaded = await storage.download(result.pieceCid)
+  const downloaded = await synapse.storage.download(result.pieceCid)
 
   // Verify
   const downloadedText = new TextDecoder().decode(downloaded)
